requirements.txt
google-api-python-client>=2.120.0
google-auth>=2.29.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
anthropic>=0.40.0
python-dotenv>=1.0.0
rich>=13.7.0
mcp>=1.0.0
markdown>=3.4.0
